+++ pait/app/sanic/_doc_route.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Optional
+from typing import Any, Optional, Type
 
 from sanic import response
 from sanic.app import Sanic
@@ -8,6 +8,7 @@ from sanic.response import HTTPResponse
 from sanic_testing.testing import SanicTestClient, TestingResponse  # type: ignore
 
 from pait.app.base.doc_route import AddDocRoute as _AddDocRoute
+from pait.app.base.doc_route import OpenAPI
 
 from ._load_app import load_app
 from ._pait import Pait
@@ -36,7 +37,7 @@ def add_doc_route(
     prefix: str = "",
     pin_code: str = "",
     title: str = "",
-    open_api_tag_list: Optional[List[Dict[str, Any]]] = None,
+    openapi: Optional[Type[OpenAPI]] = None,
     project_name: str = "",
 ) -> None:
     AddDocRoute(
@@ -45,7 +46,7 @@ def add_doc_route(
         prefix=prefix,
         pin_code=pin_code,
         title=title,
-        open_api_tag_list=open_api_tag_list,
+        openapi=openapi,
         project_name=project_name,
         app=app,
     )