+++ pait/app/starlette/_doc_route.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Optional, Set
+from typing import Any, Dict, Optional, Set, Type
 
 from starlette.applications import Starlette
 from starlette.exceptions import HTTPException
@@ -6,6 +6,7 @@ from starlette.responses import HTMLResponse, JSONResponse, Response
 from starlette.routing import Mount, Route
 
 from pait.app.base.doc_route import AddDocRoute as _AddDocRoute
+from pait.app.base.doc_route import OpenAPI
 
 from ._load_app import load_app
 from ._pait import Pait
@@ -58,7 +59,7 @@ def add_doc_route(
     prefix: str = "",
     pin_code: str = "",
     title: str = "",
-    open_api_tag_list: Optional[List[Dict[str, Any]]] = None,
+    openapi: Optional[Type[OpenAPI]] = None,
     project_name: str = "",
 ) -> None:
     AddDocRoute(
@@ -67,7 +68,7 @@ def add_doc_route(
         prefix=prefix,
         pin_code=pin_code,
         title=title,
-        open_api_tag_list=open_api_tag_list,
+        openapi=openapi,
         project_name=project_name,
         app=app,
     )