+++ pait/app/tornado/_doc_route.py
@@ -1,10 +1,11 @@
 import asyncio
 from abc import ABC
-from typing import Any, Dict, List, Optional
+from typing import Any, Optional, Type
 
 from tornado.web import Application, RequestHandler
 
 from pait.app.base.doc_route import AddDocRoute as _AddDocRoute
+from pait.app.base.doc_route import OpenAPI
 from pait.app.tornado.plugin.base import JsonProtocol
 
 from ._load_app import load_app
@@ -103,7 +104,7 @@ def add_doc_route(
     prefix: str = "",
     pin_code: str = "",
     title: str = "",
-    open_api_tag_list: Optional[List[Dict[str, Any]]] = None,
+    openapi: Optional[Type[OpenAPI]] = None,
     project_name: str = "",
 ) -> None:
     AddDocRoute(
@@ -112,7 +113,7 @@ def add_doc_route(
         prefix=prefix,
         pin_code=pin_code,
         title=title,
-        open_api_tag_list=open_api_tag_list,
+        openapi=openapi,
         project_name=project_name,
         app=app,
     )