+++ tests/test_app/test_sanic.py
@@ -14,7 +14,7 @@ from sanic_testing.testing import SanicTestClient
 from sanic_testing.testing import TestingResponse as Response  # type: ignore
 
 from example.param_verify import sanic_example
-from pait.api_doc.open_api import PaitOpenAPI
+from pait.api_doc.openapi import InfoModel, OpenAPI, ServerModel
 from pait.app import auto_load_app, get_app_attribute, set_app_attribute
 from pait.app.base.doc_route import default_doc_fn_dict
 from pait.app.sanic import TestHelper as _TestHelper
@@ -127,11 +127,11 @@ class TestSanic:
                 None,
                 str(client.get("/openapi.json?pin_code=6666")[1].text),
                 str(
-                    PaitOpenAPI(
+                    OpenAPI(
                         load_app(client.app),
-                        title="Pait Doc",
-                        open_api_server_list=[{"url": "http://localhost", "description": ""}],
-                    ).content
+                        openapi_info_model=InfoModel(title="Pait Doc"),
+                        server_model_list=[ServerModel(url="http://localhost")],
+                    ).content()
                 ),
             ).quick_ratio()
             > 0.95