+++ tests/test_app/test_tornado.py
@@ -12,7 +12,7 @@ from tornado.testing import AsyncHTTPTestCase, HTTPResponse
 from tornado.web import Application
 
 from example.param_verify import tornado_example
-from pait.api_doc.open_api import PaitOpenAPI
+from pait.api_doc.openapi import InfoModel, OpenAPI, ServerModel
 from pait.app import auto_load_app, get_app_attribute, set_app_attribute
 from pait.app.base.doc_route import default_doc_fn_dict
 from pait.app.tornado import TestHelper as _TestHelper
@@ -118,11 +118,11 @@ class TestTornado(BaseTestTornado):
             difflib.SequenceMatcher(
                 None,
                 self.fetch("/openapi.json?pin_code=6666").body.decode(),
-                PaitOpenAPI(
+                OpenAPI(
                     load_app(self._app),
-                    title="Pait Api Doc(private)",
-                    open_api_server_list=[{"url": "http://localhost", "description": ""}],
-                ).content,
+                    openapi_info_model=InfoModel(title="Pait Doc"),
+                    server_model_list=[ServerModel(url="http://localhost")],
+                ).content(),
             ).quick_ratio()
             > 0.95
         )